backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary
redis==5.0.1
stripe
python-dotenv==1.0.1
pillow==12.1.1
drf-spectacular==0.29.0
django-redis==6.0.0